+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.9.5
+
+Bug - Fix global template variables.
+
+## 0.9.4
+
+Bug - Fix query type variables.
+
+## 0.9.3
+
+Bug - Support Array data types.
+
 ## 0.9.2
 
 Bug - Fix TLS model.